Move most shared code into utils.py
